% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRelativeCoords.r
\name{calcRelativeCoords}
\alias{calcRelativeCoords}
\title{Method to calculate relative coordinates}
\usage{
calcRelativeCoords(coords, minLat, minLon)
}
\arguments{
\item{coords}{exact coordinates of a single Profile}

\item{minLat}{starting point (latititude)}

\item{minLon}{starting point (longitude)}
}
\description{
Calculates relative coordinates in meter from GPS coordinates 
(either in UTM or Gauss Krueger).
}

