% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoelectrics.r
\docType{package}
\name{geoelectrics-package}
\alias{geoelectrics}
\alias{geoelectrics-package}
\title{geoelectrics: 3D-Visualization of Geoelectric Resistivity Measurement Profiles}
\description{
Electrical resistivity tomography is an efficient geophysical technique to investigate the spatial extent of subsurface structures. Many scientific objectives in geology demand three-dimensional imaging. 
3D electrical resistivity tomography provides a technique to survey three-dimensional structures. 
Nonetheless, 3D electrical resistivity tomography requires an enormous amount of time as well as a high work load. 
In most cases, 2D electrical resistivity tomography is used to obtain two-dimensional subsurface profiles. 
This R package enables the user to visualize two-dimensional profiles in three dimensions.
}
\examples{
library(geoelectrics)
demo(geoelectrics)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kleebaum/geoelectrics}
  \item Report bugs at \url{https://github.com/kleebaum/geoelectrics/issues}
}

}
\author{
Anja Kleebaum
}
\keyword{internal}
