% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoelectricsGui.r
\name{geoelectricsGui}
\alias{geoelectricsGui}
\title{Graphical User Interface of the Geoelectrics R Package}
\usage{
geoelectricsGui()
}
\description{
This method starts a simple TclTk graphical user interface (GUI) provided by this package.
}
\examples{
geoelectricsGui()
}
