% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeightInformation.r
\docType{methods}
\name{getHeightInformation}
\alias{getHeightInformation}
\alias{getHeightInformation,ProcessedData-method}
\title{Gets the Height Information for a Profile}
\usage{
getHeightInformation(object)

\S4method{getHeightInformation}{ProcessedData}(object)
}
\arguments{
\item{object}{a single Profile.}
}
\value{
data frame containing distances and heights along the profile
}
\description{
Returns the heights for certain distances along the profile (topography information).
}
\examples{
data(sinkhole)

getHeightInformation(sinkhole@profiles[[1]]@processedData)
}
\seealso{
\code{\link{GpsCoordinates-class}}, \code{\link{Profile-class}},
\code{\link{ProcessedData-class}}
}
