% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoelectrics.r
\docType{class}
\name{GpsCoordinates-class}
\alias{GpsCoordinates-class}
\title{A class to handle gps coordinates}
\arguments{
\item{address}{address of the gps ascii file}
}
\description{
A class to handle gps coordinates
}
\section{Slots}{

\describe{
\item{\code{exact}}{data frame that contains measured gps coordinates}

\item{\code{lm}}{linear model of the measured gps coordinates}

\item{\code{relative}}{relative coordinates}

\item{\code{lmRelative}}{linear model of relative coordinates}
}}
\examples{
new("GpsCoordinates", address="../example/gps/p1.txt")
}

