% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heightAdjustment.r
\name{heightAdjustment}
\alias{heightAdjustment}
\title{Method to adjust height of a single profile}
\usage{
heightAdjustment(Profile, deltaMeter)
}
\arguments{
\item{Profile}{a single Profile}

\item{deltaMeter}{positive or negative value}
}
\description{
GPS measurements might differ otherwise.
}

