% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelplot.r
\name{levelplotLegendLabel}
\alias{levelplotLegendLabel}
\title{Levelplot Legend Label}
\usage{
levelplotLegendLabel(legend.lab = "Resistivity",
  unit = expression(paste("[", Omega, "m]")))
}
\arguments{
\item{legend.lab}{label (default: 'Resistivity').}

\item{unit}{unit (default: 'Ohm*m').}
}
\description{
Plots the label of the levelplot.
}
\examples{
data(sinkhole)

levelplot(sinkhole@profiles[[1]])
levelplotLegendLabel()

levelplot(sinkhole@profiles[[2]])
levelplotLegendLabel()

levelplot(sinkhole@profiles[[3]])
levelplotLegendLabel()
}
\seealso{
\code{\link{levelplot}}
}
