% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoelectrics.r
\name{levelplotRaw}
\alias{levelplotRaw}
\title{Plot raw data levels}
\usage{
levelplotRaw(Profile)
}
\arguments{
\item{Profile}{profile.}
}
\description{
Plots the interpolated resistance values of the
raw data without height adjustment.
}

