% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelplotXyz.r
\name{levelplotXyzHeight}
\alias{levelplotXyzHeight}
\title{Plot levels of xyz data}
\usage{
levelplotXyzHeight(Profile, xlab = "Length [m]", ylab = "Depth [m]",
  main = paste(Profile@title, "with topography"), col = colors,
  breaks = 18, trafo = log, backtrafo = exp)
}
\arguments{
\item{Profile}{profile.}

\item{xlab}{label for x-axes}

\item{ylab}{label for y-axes}

\item{main}{title to be plotted}

\item{col}{vector of colors}

\item{breaks}{number of color breaks}

\item{trafo}{transformation to be done on data (default log)}

\item{backtrafo}{back transformation to plot correct labels (default exp)}
}
\description{
Plots the interpolated resistance values of the 
xyz data after height adjustment.
}

