% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myColorRamp.r
\name{myColorRamp}
\alias{myColorRamp}
\title{Maps color to resistivity value}
\usage{
myColorRamp(col, values, minData = min(values), maxData = max(values))
}
\arguments{
\item{col}{Character vector of colors.}

\item{values}{Numeric vector of values.}

\item{minData}{Minimum value (default min(values)).
All smaller values will assigned to the first color in vector col.}

\item{maxData}{Maximum value (default max(values)).
All higher values will assigned to the last color in vector col.}
}
\description{
Maps color to (resistivity) values.
A minimum and maximum value can be specified.
}
