% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRawDataFile.r
\name{parseRawDataFile}
\alias{parseRawDataFile}
\title{Parses a Raw Data File}
\usage{
parseRawDataFile(address, skip = 9)
}
\arguments{
\item{address}{address of the raw data ascii file.}

\item{skip}{the number of lines of the data file to skip before beginning to read data.}
}
\value{
data frame containing distance, depth and resistivity values
}
\description{
Parses a geoelectrics raw data file created by the GeoTest software by Dr. Rauen.
Needs to be overwritten if another raw data format is used.
}
\examples{
fileAddress <- system.file('extdata/raw/p1_DipolDipol_SW-NE.dat',
                   package = 'geoelectrics')
                   
rawData = new('RawData')
rawData@address = fileAddress
rawData@points <- parseRawDataFile(address = fileAddress)
}
\seealso{
\code{\link{RawData-class}}, \code{\link{Profile-class}}
}
