% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoelectrics.r
\docType{methods}
\name{plot3dXyz}
\alias{Profile}
\alias{ProfileSet}
\alias{plot3dXyz}
\alias{plot3dXyz,Profile-method}
\alias{plot3dXyz,ProfileSet-method}
\alias{plot3dxyz,}
\title{Plots profiles 3D}
\usage{
plot3dXyz(.Object)

\S4method{plot3dXyz}{ProfileSet}(.Object)

\S4method{plot3dXyz}{Profile}(.Object)
}
\arguments{
\item{.Object}{either a single Profile or a ProfileSet}
}
\description{
Plots the interpolated resistance values of the 
xyz data for all profiles.
}

