% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLegend.r
\docType{methods}
\name{plotLegend}
\alias{plotLegend}
\alias{plotLegend,Profile-method}
\alias{plotLegend,ProfileSet-method}
\title{Plots the legend}
\usage{
plotLegend(.Object, legend.lab = expression(paste("Resistivity [", Omega,
  " m]")), minData = 0, maxData = 999999, breaks = NULL,
  legend.line = 2.2, nlevel = 18, lab.breaks = c(), midpoint = F,
  horizontal = T, col = colors, trafo = log, backtrafo = exp)

\S4method{plotLegend}{ProfileSet}(.Object, legend.lab,
  minData = .Object@minData, maxData = .Object@maxData)

\S4method{plotLegend}{Profile}(.Object, legend.lab,
  minData = .Object@xyzData@minData, maxData = .Object@xyzData@maxData)
}
\arguments{
\item{.Object}{either a single Profile or a ProfileSet}

\item{legend.lab}{label of legend (default: expression(paste("Resistivity [", Omega, "]")))}

\item{minData}{minimum value}

\item{maxData}{maximum value}

\item{nlevel}{number of color levels}

\item{lab.breaks}{number of breaks}

\item{col}{vector of colors}

\item{trafo}{transformation to be done on data (default log)}

\item{backtrafo}{back transformation to plot correct labels (default exp)}
}
\description{
Plots the legend for resistivity values.
}

