% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXyz.r
\name{plotXyz}
\alias{plotXyz}
\title{Plot Xyz points}
\usage{
plotXyz(Profile, xlab = "Length [m]", ylab = "Depth [m]",
  main = paste(Profile@title, "without topography"))
}
\arguments{
\item{Profile}{profile}

\item{xlab}{label for x-axes}

\item{ylab}{label for y-axes}

\item{main}{title to be plotted}
}
\description{
Plots the interpolated points of the xyz data.
}

