% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile-class.r
\docType{class}
\name{Profile-class}
\alias{Profile-class}
\title{A class to handle a single profile}
\arguments{
\item{number}{number of the profile}

\item{xyzData}{object of XyzData class}

\item{rawData}{object of rawData class}

\item{measurementType}{type of measurement (e.g. Dipole Dipole, Wenner, ...)}

\item{gpsCoordinates}{object of GpsCoordinates class}
}
\description{
A class to handle a single profile
}

