% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfileSet-class.r
\docType{class}
\name{ProfileSet-class}
\alias{ProfileSet-class}
\title{A class to handle a collection of many profiles}
\arguments{
\item{title}{title to plot}

\item{profiles}{list that contains objects of class Profile}
}
\description{
A class to handle a collection of many profiles
}
\section{Slots}{

\describe{
\item{\code{minLat}}{minimum latitude value of all profiles}

\item{\code{minLon}}{minimum longitude value of all profiles}

\item{\code{minData}}{minimum data value of all profiles}

\item{\code{maxData}}{maximum data value of all profiles}
}}

