% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawData-class.r
\docType{class}
\name{RawData-class}
\alias{RawData-class}
\title{A class to handle geoelectrics raw data}
\arguments{
\item{address}{address of the raw data ascii file}
}
\description{
A class to handle geoelectrics raw data
}
\section{Slots}{

\describe{
\item{\code{seaLevel}}{data frame that contains raw data positions and resitance values}
}}
\examples{
rawDataObject <- new("RawData", address="../example/rawdata/p1_DipolDipol_SW-NE.dat")
}

