% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo.r
\name{trafo}
\alias{trafo}
\title{Method to transform values}
\usage{
trafo(x, base = logBase)
}
\arguments{
\item{x}{values to be transformed}

\item{base}{logarithmic base}
}
\description{
Does a logarithmic transformation on the values.
}

