% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heightAdjustment.r
\name{heightAdjustment}
\alias{heightAdjustment}
\title{Adjust Profile Height}
\usage{
heightAdjustment(Profile, deltaMeter)
}
\arguments{
\item{Profile}{a single Profile.}

\item{deltaMeter}{positive or negative value.}
}
\description{
Adjusts the height of a single profile.
GPS measurement heights might differ otherwise.
}
\examples{
p3 <- new(
  "Profile",
  title = "Profile 3",
  xyzData =
    new("XyzData",
        address = system.file("extdata/processed/p3_DipolDipol_S-N.xyz", 
                  package='geoelectrics')),
  rawData =
    new("RawData",
        address = system.file("extdata/raw/p3_DipolDipol_S-N.dat", 
                  package='geoelectrics')),
  measurementType = "DipolDipol",
  gpsCoordinates =
    new("GpsCoordinates",
        address = system.file("extdata/gps/p3.txt", 
                  package='geoelectrics'))
)

p3 <- heightAdjustment(p3, -10)
}
\seealso{
\code{\link{GpsCoordinates-class}}, \code{\link{Profile-class}}
}
