% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelplotXyz.r
\name{levelplotLegendLabel}
\alias{levelplotLegendLabel}
\title{Levelplot Legend Label}
\usage{
levelplotLegendLabel(legend.lab = "Resistivity",
  unit = expression(paste("[", Omega, "m]")))
}
\arguments{
\item{legend.lab}{label (default "Resistivity").}

\item{unit}{unit (default "Ohm*m").}
}
\description{
Plots the label of the levelplot.
}

