% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoelectrics.r
\name{levelplotRaw}
\alias{levelplotRaw}
\title{Plot raw data levels}
\usage{
levelplotRaw(Profile, xlab = "Length [m]", ylab = "Depth [m]",
  main = paste(Profile@title, "without topography"))
}
\arguments{
\item{Profile}{profile.}

\item{xlab}{label for x-axes}

\item{ylab}{label for y-axes}

\item{main}{title to be plotted}
}
\description{
Plots the interpolated resistance values of the
raw data without height adjustment.
}

