% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoelectrics.r
\name{levelplotXyz}
\alias{levelplotXyz}
\title{Plot levels of xyz data}
\usage{
levelplotXyz(Profile)
}
\arguments{
\item{Profile}{profile.}
}
\description{
Plots the interpolated resistance values of the xyz 
data without height adjustment.
}

