% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoelectrics.r
\name{levelplotXyzHeight}
\alias{levelplotXyzHeight}
\title{Plot levels of xyz data}
\usage{
levelplotXyzHeight(Profile, xlab = "Length [m]", ylab = "Depth [m]",
  main = paste(Profile@title, "with topography"))
}
\arguments{
\item{Profile}{profile.}

\item{xlab}{label for x-axes}

\item{ylab}{label for y-axes}

\item{main}{title to be plotted}
}
\description{
Plots the interpolated resistance values of the 
xyz data after height adjustment.
}

