% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kleelectrics.r
\name{levelplotXyzHeight}
\alias{levelplotXyzHeight}
\title{Plot levels of xyz data}
\usage{
levelplotXyzHeight(Profile)
}
\arguments{
\item{Profile}{profile.}
}
\description{
Plots the interpolated resistance values of the 
xyz data after height adjustment.
}

