% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myColorRamp.r
\name{myColorRamp}
\alias{myColorRamp}
\title{Maps color to resistivity value}
\usage{
myColorRamp(col, values, minData, maxData)
}
\arguments{
\item{col}{vector of colors}

\item{values}{vector of values}

\item{minData}{minimum value}

\item{maxData}{maximum value}
}
\description{
Maps color to resistivity values.
}

