% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3dXyz.r
\docType{methods}
\name{plot3dXyz}
\alias{plot3d}
\alias{plot3dXyz}
\alias{plot3dXyz,Profile-method}
\alias{plot3dXyz,ProfileSet-method}
\title{Plots profiles 3D}
\usage{
plot3dXyz(.Object, title = "", sub = "", xlab = "", ylab = "",
  zlab = "", minData = 0, maxData = 9999999, col = colors,
  trafo = log)

\S4method{plot3dXyz}{ProfileSet}(.Object, title = .Object@title, sub = "",
  xlab = "", ylab = "", zlab = "", minData = .Object@minData,
  maxData = .Object@maxData, col = colors, trafo = log)

\S4method{plot3dXyz}{Profile}(.Object, title = "", sub = "", xlab = "",
  ylab = "", zlab = "", minData = .Object@xyzData@minData,
  maxData = .Object@xyzData@maxData, col = colors, trafo = log)
}
\arguments{
\item{.Object}{either a single Profile or a ProfileSet}

\item{title}{title to be plotted}

\item{sub}{subtitle to be plotted}

\item{xlab}{label of the x-axes, e.g. length [m]}

\item{ylab}{label of the y-axes, e.g. height above sea level [m]}

\item{zlab}{label of the z-axes, e.g. length [m]}

\item{col}{vector of colors}
}
\description{
Plots the interpolated resistance values of the 
xyz data for all profiles.
}

