% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLegend.r
\docType{methods}
\name{plotLegend}
\alias{plotLegend}
\alias{plotLegend,Profile-method}
\alias{plotLegend,ProfileSet-method}
\title{Plots the legend}
\usage{
plotLegend(.Object, legend.lab = expression(paste("Resistivity [", Omega,
  " m]")), minData = 0, maxData = 999999, lab.breaks = c(),
  nlevel = 128, col = colors)

\S4method{plotLegend}{ProfileSet}(.Object, legend.lab,
  minData = .Object@minData, maxData = .Object@maxData, lab.breaks, nlevel)

\S4method{plotLegend}{Profile}(.Object, legend.lab,
  minData = .Object@xyzData@minData, maxData = .Object@xyzData@maxData,
  lab.breaks, nlevel)
}
\arguments{
\item{.Object}{either a single Profile or a ProfileSet}

\item{legend.lab}{label of legend (default: expression(paste("Resistivity [", Omega, "]")))}

\item{minData}{minimum value}

\item{maxData}{maximum value}

\item{lab.breaks}{number of breaks}

\item{nlevel}{number of color levels}

\item{col}{colors}
}
\description{
Plots the legend for resistivity values.
}

