% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRaw.r
\name{plotRaw}
\alias{plotRaw}
\title{Plot raw points}
\usage{
plotRaw(Profile, xlab = "Length [m]", ylab = "Depth [m]",
  main = paste(Profile@title, "without topography"))
}
\arguments{
\item{Profile}{profile.}

\item{xlab}{label for x-axes}

\item{ylab}{label for y-axes}

\item{main}{title to be plotted}
}
\description{
Plots points of the raw data to show measurement gaps.
}

