% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kleelectrics.r
\name{plotXyzHeight}
\alias{plotXyzHeight}
\title{Plot Xyz points}
\usage{
plotXyzHeight(Profile)
}
\arguments{
\item{Profile}{profile.}
}
\description{
Plots the interpolated points of the xyz data
after height adjustment.
}

