% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile-class.r
\docType{class}
\name{Profile-class}
\alias{Profile-class}
\title{A class to handle a single profile}
\arguments{
\item{number}{number of the profile}

\item{xyzData}{object of XyzData class}

\item{rawData}{object of rawData class}

\item{measurementType}{type of measurement (e.g. Dipole Dipole, Wenner, ...)}

\item{gpsCoordinates}{object of GpsCoordinates class}
}
\description{
A class to handle a single profile
}
\examples{
new("Profile", 
     number = 1,
     xyzData = 
       new("XyzData", 
            address = "../example/xyzFiles/p1_DipolDipol_SW-NE.xyz"),
     rawData = 
       new("RawData",
            address ="../example/rawdata/p1_DipolDipol_SW-NE.dat"),
     measurementType = "DipolDipol",
     gpsCoordinates = 
       new("GpsCoordinates",
            address ="../example/gps/p1.txt"))
}

