% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.r
\docType{class}
\name{ProfileSet-class}
\alias{ProfileSet-class}
\title{Profile Set Class}
\description{
A class to handle a collection of many profiles.
}
\section{Slots}{

\describe{
\item{\code{title}}{title to plot}

\item{\code{profiles}}{list that contains objects of class Profile (\code{\link{Profile-class}})}

\item{\code{minLat}}{minimum latitude value of all profiles}

\item{\code{minLon}}{minimum longitude value of all profiles}

\item{\code{minData}}{minimum data value of all profiles}

\item{\code{maxData}}{maximum data value of all profiles}
}}

\examples{
# sinkhole <- new('ProfileSet',
#                profiles = list(p1, p2, p3),
#                title='Sinkhole')

data(sinkhole)
plot3dXyz(sinkhole)
}
\seealso{
\code{\link{Profile-class}}, \code{\link{plot3dXyz}}
}
