% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawData-class.r
\docType{class}
\name{RawData-class}
\alias{RawData-class}
\title{Raw Data Class}
\arguments{
\item{address}{address of the raw data ascii file.}
}
\description{
A class to handle geoelectrics raw data.
}
\section{Slots}{

\describe{
\item{\code{seaLevel}}{data frame that contains raw data positions and resitance values.}

\item{\code{height}}{data frame that contains topography information (distance and height).}
}}
\examples{
# rawData = new("RawData",
#                address = "../example/rawdata/p1_DipolDipol_SW-NE.dat")

data(sinkhole)
sinkhole@profiles[[2]]@rawData
sinkhole@profiles[[2]]@rawData@address
sinkhole@profiles[[2]]@rawData@height
sinkhole@profiles[[2]]@rawData@seaLevel
}
\seealso{
\code{\link{Profile-class}}, \code{\link{ProfileSet-class}}
}

