\name{sinkhole}
\docType{data}
\alias{sinkhole}
\title{Filled Sinkhole}
\description{Geoelectrics profiles measured at a filled sinkhole.
This data set contains an object of the ProfileSet class.
}
\usage{
# data(sinkhole)
# plot3dXyz(sinkhole)
# plotLegend(sinkhole)

# plotIntersect(sinkhole)

# levelplotXyz(sinkhole@profiles[[1]])
# levelplotLegendLabel()

# plotRaw(sinkhole@profiles[[2]])
# plotRawHeight(sinkhole@profiles[[2]])

# levelplotRaw(sinkhole@profiles[[2]])
# levelplotLegendLabel()
}
\format{Object of Profile Set class including three Profiles.}
\keyword{datasets}
