% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoelectrics.r
\docType{class}
\name{XyzData-class}
\alias{XyzData-class}
\title{A class to handle xyz data}
\arguments{
\item{address}{address of the xyz ascii file}
}
\description{
The software Res2DInv produces .xyz-files that contain the
inverted resistance values. This class parses this kind of files.
}
\section{Slots}{

\describe{
\item{\code{seaLevel}}{data frame that contains positions and values withouth height adjustment}

\item{\code{heightAdaption}}{data frame that contains positions and values after height adjustment}
}}
\examples{
new("XyzData", address="../example/xyzFiles/p1_DipolDipol_SW-NE.xyz")
}

